Move you to the top
I hope you don't mind being first since my instructions suggest that new students go to the head of the class.

Thanks for starting the discussion about your M1. That's another way to get people involved in this community and I might be able to use that as content elsewhere.